Extract Sales Assist engine
